refactor: Fine tuning the const
This will cause a huge improvement on the execution time of the tests.

Test does not seems to fail, apparently it works very good.

Tests:

avatar from 9s to 4.5s
authcode from 20-30 seconds -> 11.5
accordion from seconds 41s-> 25s

This is just a quality of life commit finetuning for now, to save 30 seconds each time we run the tests locally during the development.